python-sdk: split convenience helpers out of api
